requirements.txt
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
pillow==10.1.0
transformers==4.35.2
tokenizers==0.15.0
python-decouple==3.8
accelerate==0.24.1
